requirements.txt
flask
sentence-transformers
faiss-cpu
transformers